requirements.txt
aiomysql==0.2.0
asgiref==3.8.1
fastapi==0.111.1
fastapi-cli==0.0.4
Flask==3.0.3
gunicorn==23.0.0
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
requests==2.32.3
uuid==1.30
uvicorn==0.30.3
wheel==0.44.0
scikit-learn==1.3.2
prisma==0.15.0
nodeenv==1.9.1
tomlkit==0.13.2
psycopg2-binary==2.9.9